* DonDon.csproj: * Utility.cs: * Settings.cs: * NetworkHelper.cs: * LoginActivity.cs: * OrderActivity.cs: * Home.axml: * Order.axml: * LoginController.cs: * OrderList.cs: * OrderController.cs: * AndroidManifest.xml: * OrderList1.cs: * Resource.designer.cs: * OrderAddListAdapter.cs: * Home.axml: * Order.axml: * Login.axml: * RestaurantSelectActivity.cs: * Home.axml: * Order.axml: * Login.axml:
* HomeActivity.cs: 123

+++ Adapters/OrderAddListAdapter.cs
@@ -49,12 +49,12 @@ namespace DonDon
 			return _OrderList;
 		}
 
-		public void SetStockAtPosition(int position, int StockNumber)
+		public void SetStockAtPosition(int position, double StockNumber)
 		{
 			_OrderList[position].StockNumber = StockNumber;
 		}
 
-		public void SetOrderAtPosition(int position, int OrderNumber)
+		public void SetOrderAtPosition(int position, double OrderNumber)
 		{
 			_OrderList[position].OrderNumber = OrderNumber;
 		}
